# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reports/academic": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create Academic Report",
        "operationId": "create_academic_report_api_v1_reports_academic_post",
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicReportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "List My Reports",
        "operationId": "list_my_reports_api_v1_reports_academic_get",
        "security": [
          {
            "Access Token": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicReportListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/academic/{year}/{month}": {
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "Get Report By Year Month",
        "operationId": "get_report_by_year_month_api_v1_reports_academic__year___month__get",
        "security": [
          {
            "Access Token": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicReportLookupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/academic/{report_id}": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update Academic Report",
        "description": "Update a draft academic report. Cannot edit after final submission.",
        "operationId": "update_academic_report_api_v1_reports_academic__report_id__patch",
        "security": [
          {
            "Access Token": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicReportUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/academic/{report_id}/submit": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Submit Academic Report",
        "description": "Finalize and submit an academic report. Cannot be undone.",
        "operationId": "submit_academic_report_api_v1_reports_academic__report_id__submit_post",
        "security": [
          {
            "Access Token": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/academic/admin/users/{user_id}/monitoring/{year}": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Enable Academic Monitoring",
        "description": "Enable academic monitoring for a user for a specific year.",
        "operationId": "enable_academic_monitoring_api_v1_reports_academic_admin_users__user_id__monitoring__year__post",
        "security": [
          {
            "Access Token": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "academic"
        ],
        "summary": "Disable Academic Monitoring",
        "description": "Disable academic monitoring for a user for a specific year.",
        "operationId": "disable_academic_monitoring_api_v1_reports_academic_admin_users__user_id__monitoring__year__delete",
        "security": [
          {
            "Access Token": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/academic/admin/users/{user_id}/monitoring": {
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "Get User Monitoring Years",
        "description": "Get all years a user is monitored for.",
        "operationId": "get_user_monitoring_years_api_v1_reports_academic_admin_users__user_id__monitoring_get",
        "security": [
          {
            "Access Token": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/academic/admin/users/{user_id}": {
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "List User Reports",
        "operationId": "list_user_reports_api_v1_reports_academic_admin_users__user_id__get",
        "security": [
          {
            "Access Token": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicReportListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AcademicReportCreate": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/GoalCreate"
            },
            "type": "array",
            "minItems": 2,
            "title": "Goals"
          },
          "evidence_urls": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Urls"
          }
        },
        "type": "object",
        "required": [
          "year",
          "month",
          "goals"
        ],
        "title": "AcademicReportCreate"
      },
      "AcademicReportListResponse": {
        "properties": {
          "reports": {
            "items": {
              "$ref": "#/components/schemas/AcademicReportResponse"
            },
            "type": "array",
            "title": "Reports"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "reports",
          "total"
        ],
        "title": "AcademicReportListResponse"
      },
      "AcademicReportLookupResponse": {
        "properties": {
          "exists": {
            "type": "boolean",
            "title": "Exists"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AcademicReportResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "exists"
        ],
        "title": "AcademicReportLookupResponse"
      },
      "AcademicReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "month": {
            "type": "integer",
            "title": "Month"
          },
          "is_submitted": {
            "type": "boolean",
            "title": "Is Submitted"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "evidence_urls": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Urls"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/GoalResponse"
            },
            "type": "array",
            "title": "Goals"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "year",
          "month",
          "is_submitted",
          "created_at",
          "submitted_at",
          "evidence_urls",
          "goals"
        ],
        "title": "AcademicReportResponse"
      },
...
      "AcademicReportUpdate": {
        "properties": {
          "goals": {
            "items": {
              "$ref": "#/components/schemas/GoalCreate"
            },
            "type": "array",
            "minItems": 2,
            "title": "Goals"
          },
          "evidence_urls": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Urls"
          }
        },
        "type": "object",
        "required": [
          "goals"
        ],
        "title": "AcademicReportUpdate"
      },
...
      "GoalCreate": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/AcademicGoalCategory"
          },
          "custom_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Category"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "achievement_pct": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Achievement Pct"
          }
        },
        "type": "object",
        "required": [
          "category",
          "content"
        ],
        "title": "GoalCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "Access Token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}